Remove the exercise/solution setup (for now), and enable the tests for what were the solution classes.
