+++ statefun-workshop-functions/src/main/java/com/ververica/statefun/workshop/provider/MerchantProvider.java
@@ -16,7 +16,7 @@
 
 package com.ververica.statefun.workshop.provider;
 
-import com.ververica.statefun.workshop.functions.solutions.MerchantFunction;
+import com.ververica.statefun.workshop.functions.MerchantFunction;
 import com.ververica.statefun.workshop.utils.MerchantScoreService;
 import com.ververica.statefun.workshop.utils.ProductionMerchantScoreService;
 import org.apache.flink.statefun.sdk.FunctionType;
