+++ statefun-workshop-functions/src/test/java/com/ververica/statefun/workshop/functions/FraudCountTest.java
@@ -20,7 +20,6 @@ import static com.ververica.statefun.workshop.identifiers.FRAUD_FN;
 import static org.apache.flink.statefun.testutils.matchers.StatefulFunctionMatchers.*;
 import static org.hamcrest.Matchers.equalTo;
 
-import com.ververica.statefun.workshop.functions.exercises.FraudCount;
 import com.ververica.statefun.workshop.generated.ConfirmFraud;
 import com.ververica.statefun.workshop.generated.QueryFraud;
 import com.ververica.statefun.workshop.generated.ReportedFraud;
@@ -37,7 +36,6 @@ public class FraudCountTest {
     private static final Address SENDER = new Address(new FunctionType("ververica", "sender"), "id");
 
     @Test
-    @Ignore
     public void testCounting() {
         FunctionTestHarness harness = FunctionTestHarness.test(ignore -> new FraudCount(), FRAUD_FN, "user1");
 
@@ -53,7 +51,6 @@ public class FraudCountTest {
     }
 
     @Test
-    @Ignore
     public void testRollingCount() {
         FunctionTestHarness harness = FunctionTestHarness.test(ignore -> new FraudCount(), FRAUD_FN, "user1");
