+++ statefun-workshop-functions/src/test/java/com/ververica/statefun/workshop/functions/TransactionManagerTest.java
@@ -24,7 +24,6 @@ import static org.hamcrest.Matchers.equalTo;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.Timestamp;
-import com.ververica.statefun.workshop.functions.exercises.TransactionManager;
 import com.ververica.statefun.workshop.generated.FeatureVector;
 import com.ververica.statefun.workshop.generated.FraudScore;
 import com.ververica.statefun.workshop.generated.QueryFraud;
@@ -38,7 +37,6 @@ import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
 
-@Ignore
 public class TransactionManagerTest {
 
     private static final String ACCOUNT = "account-id";
